refactor: switch config siblings to self-reexport imports
