+++ packages/opencode/src/cli/cmd/tui/config/tui.ts
@@ -10,8 +10,8 @@ import { Global } from "@/global"
 import { AppFileSystem } from "@opencode-ai/shared/filesystem"
 import { Npm } from "@opencode-ai/shared/npm"
 import { CurrentWorkingDirectory } from "./cwd"
-import { ConfigPlugin } from "@/config"
-import { ConfigKeybinds } from "@/config"
+import { ConfigPlugin } from "@/config/plugin"
+import { ConfigKeybinds } from "@/config/keybinds"
 import { InstallationLocal, InstallationVersion } from "@/installation/version"
 import { makeRuntime } from "@/cli/effect/runtime"
 import { Filesystem, Log } from "@/util"
