+++ packages/opencode/src/config/index.ts
@@ -1,5 +1,3 @@
 export * as Config from "./config"
 export * as ConfigMarkdown from "./markdown"
 export * as ConfigPaths from "./paths"
-export * as ConfigKeybinds from "./keybinds"
-export * as ConfigPlugin from "./plugin"
